Week 7 exercises.

+++ app/helpers/application_helper.rb
@@ -3,18 +3,17 @@ module ApplicationHelper
     unless current_user.nil?
       edit = link_to('Edit', url_for([:edit, item]), class: "btn btn-primary")
       if current_user.admin == true
-        del = link_to('Destroy', item, method: :delete,
-                                    data: { confirm: 'Are you sure?' },
-                                    class: "btn btn-danger")
+        del = link_to('Destroy', item, method: :delete, data: { confirm: 'Are you sure?' }, class: "btn btn-danger")
       end
       raw("#{edit} #{del}")
     end
   end
 
   def round(number)
-    if(number == nil)
+    if number.nil?
 
       return 0
+
     end
     number_with_precision(number, precision: 1)
   end
